requirements.txt
Django==4.2.2
djangorestframework==3.14.0

ipython==8.14.0
python-dotenv==1.0.0
drf-yasg==1.21.6
PyYAML==6.0
pandas==2.0.3

isort==5.12.0
flake8==6.0.0
flake8-absolute-import==1.0.0.1
flake8-annotations==3.0.1
flake8-bugbear==23.6.5
flake8-builtins==2.1.0
flake8-comprehensions==3.13.0
flake8-django==1.3
flake8-pep3101==2.0.0
flake8_simplify==0.20.0
pep8-naming==0.13.3